Clean up command handler

+++ src/handlers/CommandHandler.ts
@@ -76,12 +76,9 @@ class CommandHandler{
 			// Starting at 1 index so that it takes away the prefix
 			// This makes it easier to later allow custom prefixes for servers, and just check for those too in the if case above
 			args[0] = args[0].substring(config.bot.defaultPrefix.length);
+			console.log(args);
 			this.moduleList.forEach(async (module) => {
 				if (!message.member){
-					return;
-				}
-				if (!module.checkPermissions(message.member.permission)){
-					message.channel.createMessage("You don't have permission to use this command");
 					return;
 				}
 				if (module.name.toLowerCase() === "owner"){
@@ -96,12 +93,15 @@ class CommandHandler{
 						message.channel.createMessage("You don't have permission to use this command");
 						return;
 					}
+					if (!module.checkPermissions(message.member.permission)){
+						message.channel.createMessage("You don't have permission to use this command");
+						return;
+					}
 
 					if (command.deleteCommand === true){
 						await message.delete();
 					}
 
-					args.shift();
 					await command.commandFunc(message, args, this.db);
 
 					return true;
